Refactor event structure and implement level
