+++ src/main/java/de/hdg/keklist/Keklist.java
@@ -12,6 +12,11 @@ import de.hdg.keklist.commands.KeklistCommand;
 import de.hdg.keklist.commands.WhitelistCommand;
 import de.hdg.keklist.database.DB;
 import de.hdg.keklist.events.*;
+import de.hdg.keklist.events.feats.ListPingEvent;
+import de.hdg.keklist.events.feats.NotifyJoinEvent;
+import de.hdg.keklist.events.qol.BlacklistRemoveMotd;
+import de.hdg.keklist.events.qol.ListCommandPageEvent;
+import de.hdg.keklist.events.qol.ServerWhitelistChangeEvent;
 import de.hdg.keklist.extentions.GeyserEventRegistrar;
 import de.hdg.keklist.extentions.PlaceholderAPIExtension;
 import de.hdg.keklist.extentions.context.BlacklistedCalculator;
