+++ src/main/java/de/hdg/keklist/events/feats/GeyserConnectionEvent.java
@@ -1,4 +1,4 @@
-package de.hdg.keklist.events;
+package de.hdg.keklist.events.feats;
 
 import de.hdg.keklist.Keklist;
 import de.hdg.keklist.extentions.WebhookManager;
