+++ src/main/java/de/hdg/keklist/events/qol/BlacklistRemoveMotd.java
@@ -1,4 +1,4 @@
-package de.hdg.keklist.events;
+package de.hdg.keklist.events.qol;
 
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
