+++ src/main/resources/config.yml
@@ -93,6 +93,7 @@ messages:
       - "<red><bold>Sorry, but you're not allowed to join this server!"
 
     whitelisted:
+      - level: "<red>Your whitelist level is insufficient, thus you can not join!"
       - "<gold><bold>You're kicked because the server is in whitelist mode!"
       - "<gold><bold>Sorry, but you're not on the whitelist!"
